backend/requirements.txt
fastapi>=0.115.0
uvicorn[standard]>=0.30.0
pydantic>=2.8.0
pydantic-settings>=2.4.0
python-dotenv>=1.0.1

supabase>=2.6.0
httpx>=0.27.0
tenacity>=8.5.0

scrapling[fetchers]>=0.2.99

beautifulsoup4>=4.12.3
lxml>=5.3.0
tldextract>=5.1.2

dnspython>=2.6.1

twilio>=9.2.3

google-api-python-client>=2.141.0
google-auth>=2.33.0
google-auth-httplib2>=0.2.0
google-auth-oauthlib>=1.2.1

apscheduler>=3.10.4
